changed compiler to cc, fixed split, strnstr, substr compiler errors

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: adapassa <adapassa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/27 17:45:44 by adapassa          #+#    #+#             */
-/*   Updated: 2024/01/03 20:47:34 by adapassa         ###   ########.fr       */
+/*   Updated: 2024/01/08 15:14:12 by adapassa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,8 @@
 
 char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if ((!*big && !*little) || !*little)
